refactor(windows): finestre spostate nella directory windows
